requirements.txt
httpx>=0.27.0
pydantic>=2.6.0
pyyaml>=6.0.1
matplotlib>=3.8.0
plotly>=5.18.0
rich>=13.7.0
python-dotenv>=1.0.0
numpy>=1.26.0
kaleido>=0.2.1
json-repair>=0.30.0
